change fnv1a return 64bit

+++ c_src/hm_fnv.c
@@ -7,7 +7,7 @@ static uint64_t FNV_64_PRIME = UINT64_C(0x100000001b3);
 static uint32_t FNV_32_INIT = 2166136261UL;
 static uint32_t FNV_32_PRIME = 16777619;
 
-uint32_t hash_fnv1_64(const char *key, size_t key_length){
+uint64_t hash_fnv1_64(const char *key, size_t key_length){
 
     uint64_t hash = FNV_64_INIT;
     size_t x;
@@ -17,17 +17,17 @@ uint32_t hash_fnv1_64(const char *key, size_t key_length){
       hash ^= (uint64_t)key[x];
     }
 
-    return (uint32_t)hash;
+    return hash;
 }
 
-uint32_t hash_fnv1a_64(const char *key, size_t key_length){
-    uint32_t hash = (uint32_t) FNV_64_INIT;
+uint64_t hash_fnv1a_64(const char *key, size_t key_length){
+    uint64_t hash = (uint64_t) FNV_64_INIT;
     size_t x;
 
     for (x = 0; x < key_length; x++) {
-      uint32_t val = (uint32_t)key[x];
+      uint64_t val = (uint64_t)key[x];
       hash ^= val;
-      hash *= (uint32_t) FNV_64_PRIME;
+      hash *=  FNV_64_PRIME;
     }
 
     return hash;
